refactor progress bar
