+++ src/main/steps/common/task-list/components/notification-banner/config/da_applicant.ts
@@ -1,8 +1,7 @@
-import { CaseWithId } from "../../../../../../app/case/case";
-import { BannerNotification, notificationBanner } from "../utils";
-import { UserDetails } from "../../../../../../app/controller/AppRequest";
-import { State, YesOrNo } from "../../../../../../app/case/definition";
-
+import { CaseWithId } from '../../../../../../app/case/case';
+import { State, YesOrNo } from '../../../../../../app/case/definition';
+import { UserDetails } from '../../../../../../app/controller/AppRequest';
+import { BannerNotification, notificationBanner } from '../utils';
 
 export const DA_APPLICANT = [
   {
@@ -27,4 +26,4 @@ export const DA_APPLICANT = [
       return caseData?.state === State.CASE_CLOSED;
     },
   },
-]
+];
