+++ src/main/steps/common/task-list/components/progress-bar/config/index.ts
@@ -1,0 +1,5 @@
+import { CA_APPLICANT } from './ca_applicant';
+import { CA_RESPONDENT } from './ca_respondent';
+import { DA_APPLICANT } from './da_applicant';
+import { DA_RESPONDENT } from './da_respondent';
+export default { CA_APPLICANT, DA_APPLICANT, CA_RESPONDENT, DA_RESPONDENT };
