+++ src/main/steps/prl-cases/dashboard/content.test.ts
@@ -150,7 +150,7 @@ describe('Dashboard content', () => {
           rows: [
             [
               {
-                html: '<a class=\"govuk-link\" href=\"/case/1675347915490145\">1675347915490145</a>',
+                html: '<a class="govuk-link" href="/case/1675347915490145">1675347915490145</a>',
               },
               {
                 text: 'FL401',
