+++ src/main/steps/prl-cases/dashboard/tabView.test.ts
@@ -164,7 +164,7 @@ describe('Dashboard tab content', () => {
           rows: [
             [
               {
-                html: '<a class=\"govuk-link\" href=\"/case/1675347915490145\">1675347915490145</a>',
+                html: '<a class="govuk-link" href="/case/1675347915490145">1675347915490145</a>',
               },
               {
                 text: 'FL401',
