+++ src/main/steps/prl-cases/dashboard/tabView.ts
@@ -4,11 +4,7 @@ import { CaseWithId } from '../../../app/case/case';
 import { CaseType, PartyType, State } from '../../../app/case/definition';
 import { isCaseLinked } from '../../../steps/common/task-list/utils';
 import { applyParms } from '../../common/url-parser';
-import {
-  C100_RETRIVE_CASE,
-  FETCH_CASE_DETAILS,
-  PageLink,
-} from '../../urls';
+import { C100_RETRIVE_CASE, FETCH_CASE_DETAILS, PageLink } from '../../urls';
 
 import { UserDetails } from './../../../app/controller/AppRequest';
 import { getCasePartyType } from './utils';
@@ -270,7 +266,7 @@ const getTaskListUrl = (
   let url;
 
   if (linkPartyType === PartyType.RESPONDENT) {
-      url = applyParms(`${FETCH_CASE_DETAILS}`, { caseId: caseNumber });
+    url = applyParms(`${FETCH_CASE_DETAILS}`, { caseId: caseNumber });
   } else {
     if (caseType === CaseType.C100) {
       if (State.AWAITING_SUBMISSION_TO_HMCTS === caseStatus) {
@@ -279,7 +275,7 @@ const getTaskListUrl = (
         url = applyParms(`${FETCH_CASE_DETAILS}`, { caseId: caseNumber });
       }
     } else {
-      url = applyParms(`${FETCH_CASE_DETAILS}`, { caseId: caseNumber });;
+      url = applyParms(`${FETCH_CASE_DETAILS}`, { caseId: caseNumber });
     }
   }
