+++ src/main/steps/urls.ts
@@ -622,7 +622,7 @@ export const SAFEGAURD_EXCLUDE_URLS = [
   APPLICANT_CA_DA_REQUEST,
   FETCH_CASE_DETAILS,
   PIN_ACTIVATION_CASE_ACTIVATED_URL,
- // RESPONDENT_TASK_LIST_URL,
+  // RESPONDENT_TASK_LIST_URL,
   DOCUMENT_MANAGER,
   CITIZEN_DOWNLOAD_UPLOADED_DOCS,
 ];
